Report Viper Consistency Errors (#41)
Note that in theory this consistency check should be done automatically
when verifying the program. However, since this is not the case, we
insert a manual check to catch these errors.
This also means that the nullable test case now fails as the conversion
from non-nullables to nullables is not handles correctly.