Add a command-line processor to the compiler.
Allows enabling or disabling the VIPER_TEXT info message, which is
not supposed to be shown to end users.

The `getProgramForLogging` function may seem overcomplicated; I've
implemented it this way to allow for a more succinct Viper dump down the
line, where we don't include standard things like domains and fields.