Add a `Result` notion to simplify complex visits
Now when a visit requires generating a temporary variable for the
result, we can do this via the context instead of manually threading the
variable (e.g. in `convertWhenBranches`).

Note that it is now best practice to use the `convert` function of the
context to convert subexpressions, rather than calling `accept`.