Fix recursive class issue.
Our converter currently goes into a loop when it encounters a recursive
class (one with a field with a type that uses the class' own type).

We can fix this by splitting up the data a bit; not sure this is the
best solution in the long term, but we can find a more elegant one as
our class embedding gets more complicated.