Change try-catch implementation to match Kotlin's CFG
In Kotlin, the semantics of the CFG are that a try block may go to the
catch block at the beginning or the end of the block
(nondeterministically).  This gives slightly different verification
results than what we got in our implementation.  From discussions with
the Language Evolution team, the default behaviour is the desired one,
so I'm switching the implementation to this simpler one so that we get
verification failures in the same situations as they do.

More details to be found here: https://kotlinlang.org/spec/control--and-data-flow-analysis.html#other-expressions