Add comparison operators (#132)
Comparison in Kotlin are implemented in the following way: `a < b`
corresponds to `a.compareTo(b) < 0`
The `compareTo` method is nothing more than a subtraction. I don't know
if this method deserves to be represented as a `specialKotlinFunction`
since I think it is not used much. Moreover it has several overloadings
that can make the encoding a bit hard