Do position handling via `LinearizationContext` (#160)
This is the first behaviour-changing PR of a migration to linearization
post FIR -> `ExpEmbedding` conversion. Where previously position
information was embedded in each `ExpEmbedding` node, we now have nodes
that set position information in each subtree.

At the moment, this only uses the pure linearizer. The next step is to
use the real linearizer to perform some part of the translation, but it
is not clear how this could be done gradually; it seems like we may need
to have some kind of `StmtConverter`-as-`Linearizer` step to begin with,
where we generate `ExpEmbedding`s and then very soon linearize them,
just to be able to gradually move everything into the `ExpEmbedding`
structure.