Add error warning messages for lambdas that may leak from the `callsInPlace` effect. (#184)
The commit introduces a new `SourceRole` called
`ParamFunctionLeakageCheck` that is embedded as info metadata in
`special$duplicable` function calls. When a `PreconditionInCallFalse` or
`AssertFailed` error is raised by Viper, and it contains the new defined
role, then we are dealing with a possible lambda leakage. In addition,
to output the parameter is leaking, now the `FirVariableEmbedding`, when
converted into Viper's code, it embeds the symbol as info metadata.
Therefore, during error reporting, we are now able to output a
user-friendly warning message.

> Note: in the next PR I will refactor the code of
`VerifierErrorInterpreter` since the cyclomatic complexity of
`reportVerificationErrorUserFriendly` is growing.