Properties Modernization (#208)
This pull request
1. removes receiver dependent name mangling for non-private properties;
2. fixes bugs related to open fields and complicated cases of overriding
(`val` -> `var`, different types);
3. "immutable" predicates are inhaled rather than being required.

The field is used for property iff it is final and has
`FirDefaultPropertyGetter`, `FirDefaultPropertySetter` or both.
This pull request does not try to use fields when type of receiver is
final (and property itself is not) although it is a possible further
step.

You might want to take a look at new tests:
`good_contracts/multiple_interfaces`,
`no_contracts/multiple_interfaces`, `good_contracts/private_properties`.


```kotlin
interface Base
fun takeBase(base: Base) { }

class Impl : Base
fun createImpl() {
  val impl = Impl()
  takeBase(impl)
}
```
Note that 3. is not a final solution, but rather a workaround. It is
needed because otherwise the example above won't have correct
permissions. In case we want to add mutable fields with `write` accesses
to this predicate, we won't be able to inhale permissions to them twice.
Places where changes have been made are marked with `//TODO (inhale vs
require)`.