Split class details into separate object. (#224)
Right now, details like supertypes and properties are all included in
the `ClassTypeEmbedding`. This is unfortunate, because these are data
classes, and this breaks the equality definition on them. With this
commit, all the mutable data is split off into a separate object. The
class type embeddings still aren't quite interchangable, since some may
be missing this `details` object, but at least it's much easier to
detect now.