Nullable type embedding name fix (#239)
We didn't put `'T'` tag to the nullable type names apparently and only
added `'N'` prefix when scope name was not null which sounds very
strange. I tried to configure a more plausible name for that case, but
I'm still open to suggestions (collisions are pretty easy to find
currently in function names).
Also added a test when functions only differ in the nullability of their
argument (it's allowed on JVM).