Reorganization of tests in good_contracts (#242)
Currently we have plenty of tests which should be verified but don't
contain any contracts and/or don't check any properties related to
translation of contracts. Hence, a broader `verifies` directory is
introduced instead of `good_contracts` with tests divided by topic.

Where possible, contracts are replaced with `verify` builtin. It is also
made `vararg`.