Correct Invariants for loops (#246)
With this PR, we capture all the locals and parameters that can be seen
in the loop in its invariant in order to guarantee a proper access to
the fields of the variables. Invariants we inserted before did not allow
this.

```
while (df$rt$boolFromRef(anon$6))
    invariant df$rt$isSubtype(df$rt$typeOf(l2$iteration), df$rt$intType())
    invariant acc(p$c$ClassWithField$shared(l0$local), wildcard)
    invariant df$rt$isSubtype(df$rt$typeOf(l0$local), df$rt$c$ClassWithField())
{
    ...
}
```

Currently we include all possible variables here.
In the future, we would like to

- only leave the variables that are used in the loop;
- inhale invariants based on the information which can be deduced by
type of `ExpEmbedding` when we enter the loop (e.g. when the variable
should be smartcast)