Support preconditions in functions (#250)
This PR enables preconditions in functions, like this:
```kotlin
fun foo(idx: Int) {
    preconditions {
        0 <= idx
        idx < 4
    }
}
```

All the preconditions must coexist in the single block in the beginning
of the function, otherwise they won't be processed.

Note that current implementation conflicts with inlining (preconditions
of the inlined function won't be used).
This PR also unifies the ways we handle `loopInvariants`,
`preconditions` and `postconditions`.