User-specified postconditions (#252)
This pull request introduces user-specified postconditions.

A single `postconditions` scope is allowed, either in the very beginning
of the function or right after `preconditions` scope.
Note that `postconditions` consumes generic function, but the type in
the lambda must always coincide with the return type of the function.

Example:
```kotlin
fun decreaseByTen(int: Int) {
    preconditions {
        int >= 10
    }
    postconditions<Int> { res ->
        res >= 0
    }
    return int - 10
}
```