Encapsulate callables as `CallableEmbedding` (#101)
The goal of this change is to make callable objects responsible for the
calls they make, so that calling code doesn't have to worry about how
the call should be handled.

This commit does not yet handle lambdas; those seem like they are likely
to be more complicated, but should eventually also be subsumed into this
structure.

Eventually, we'd also like calls to all be via `ExpEmbedding`s, but this
first requires lambdas to be represented as `ExpEmbedding`s as well.