Prevent duplicate `NullableType` application. (#142)
In `null_cast`, we previously applied `NullableType` to an arbitrary
`newType : Type`. This could lead to a type getting double nullability,
which can cause issues in reasoning as it does not make sense in Kotlin.
With this change, we instead check that the `newType` is indeed nullable
before permitting reasoning about the cast.