Update invariants and pre-/postconditions to use `ExpEmbedding` (#157)
Previously, we directly produced `Exp` in these constructions, but this
is inconvenient as the rest of our code gradually becomes increasingly
high-level. This change updates the interface of invariants and
pre-/postconditions to use `ExpEmbedding` and thus brings it in line
with the rest of our code. Under the hood, we still occasionally need to
generate `Exp` instances and then wrap them, but this is something that
can be tackled gradually.

This change is in particular an important step towards moving the
linearization step to after the conversion from FIR to `ExpEmbedding`.