Implement statement modifier system for linearisation (#213)
Previously, we have dealt with `Inhale` and `Exhale` statements by
explicitly inserting them around statements that required them. This
change makes these statements modifiers on the statement they refer to;
this should open the room for increased flexibility, and in particular
make it easier to implement an unfolding/folding system.