Add a builder for `TypeEmbedding`s (#229)
Previously, we used the type embeddings directly in the code. This makes
changing their representation quite hard; this PR introduces a builder
that lets us hide the implementation details.

This isn't quite perfect; we sometimes still check what subtype a
particular embedding is. Getting rid of that entirely would be an even
bigger change, though.