Split `TypeEmbedding` into a `PretypeEmbedding` and flags. (#241)
This makes it easier to uniformly track flags for our embeddings; in
Kotlin this is normally just nullability, but we also need to track
uniqueness and borrowing.

This PR just restructures the code, it doesn't add any new support for
unique/borrowing yet.

The PR does not attempt to fix the naming issues; it would be good to
structure naming better, but I'll do that in a follow-up PR.